requirements.txt
streamlit>=1.32.0
pydantic>=2.6.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pandas>=2.2.0
openpyxl>=3.1.2
python-docx>=1.1.0
pdfplumber>=0.11.0
Pillow>=10.0.0
